fix: ビルドエラーをすべて解消
- H264Decoder: MFTEnumEx をvoid+outパターンに修正、EnumFlagを生値に変更
- H264Decoder: MFCreateDXGIDeviceManager/ResetDevice シグネチャ修正
- H264Decoder: OutputStreamInfoFlags.ProvidesSamples を 0x100 リテラルに変更
- MonitorInfo: Index・RefreshRate プロパティを追加、EnumDisplaySettings で取得
- BridgeConfig.InputConfig: MacDisplayWidth / MacDisplayHeight フィールドを追加
- BridgeOrchestrator: DiscoverMacAsync 引数修正、GlobalInputHook を追加
- BridgeOrchestrator: SmartKvmService コンストラクタ・Start・DisposeAsync を修正

+++ src/Windows/MacWinBridge.Display/Rendering/FullScreenRenderer.cs
@@ -68,8 +68,8 @@ public sealed class FullScreenRenderer : IDisposable
 
         var desc = new SwapChainDescription1
         {
-            Width       = (uint)_targetWidth,
-            Height      = (uint)_targetHeight,
+            Width       = _targetWidth,
+            Height      = _targetHeight,
             Format      = Format.B8G8R8A8_UNorm,
             Stereo      = false,
             SampleDescription = new SampleDescription(1, 0),
